+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.5.0
 
+* [WARNING] `Theme.of(context)` is now used to determine text styles
+* title and description can now be null
+* added option to disable pulsing animation
 * added possibility to pass any iterable containing feature ids to `FeatureDiscovery.discoverFeatures`
 * added the `@required` annotation to parameters that cannot be null